This commit includes a few backwards-compatibility breaking changes.
First of all the database models changed a bit, so you'll need to rebuild the
reference database.

Secondly, increased exposure to DRF best practice has made me realise, we should
not be using PUT to partially modify objects. The contract for PUT is to
replace completely the contents of a given resource with new content from the
client. We should be using PATCH for partial updates.

So I've changed the way to partially update an Alloc Rule to accept PATCH and
to not offer PUT. I intend to also soon backtrack on using PUT to add a new
unspecified rule on a Rule List.  Watch this space.

I've had a change of heart on the documentation that comes in the browsable API
headers as well. I've realised, that if I enter fully into the spirit of REST
best practice, we get all that is needed auto-generated, apart from background
info - which I'll retain.

I add validation on the fields of an Alloc Rule.

In answer to the question - what payload types are legal? - The only way I can
publish that from the API is through the end point documentation - which I'll
add.